Fixed code duplication in some functions with songById/playlistById. Also changed closest element and time in searchByDuration
Instead of using songById/playlistById twice in some funcitons (one for valid id check, other for actual use), assign the function value to a variable- thus checking for a valid id and also saving the value for later usage.
searchByDuration- instead of assigning a value from one of the arrays to closest time and element, and then having to make sure the arrays aren't empty, both variables are now undefined, and the "if" statements now take care of an undefined closestElement variable situation.